Revert "Move global metric back to `otel/metric/global` for minor release (#3986)"
This reverts commit 8dba38e02f16d29a315d2e6890d21f763677831a.
