+++ go.mod
@@ -7,6 +7,7 @@ require (
 	github.com/go-logr/stdr v1.2.2
 	github.com/google/go-cmp v0.5.9
 	github.com/stretchr/testify v1.8.2
+	go.opentelemetry.io/otel/metric v0.38.0
 	go.opentelemetry.io/otel/trace v1.15.0
 )
