+++ internal/global/instruments.go
@@ -12,13 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package global // import "go.opentelemetry.io/otel/metric/internal/global"
+package global // import "go.opentelemetry.io/otel/internal/global"
 
 import (
 	"context"
 	"sync/atomic"
 
-	"go.opentelemetry.io/otel"
 	"go.opentelemetry.io/otel/metric"
 	"go.opentelemetry.io/otel/metric/embedded"
 )
@@ -44,7 +43,7 @@ var _ metric.Float64ObservableCounter = (*afCounter)(nil)
 func (i *afCounter) setDelegate(m metric.Meter) {
 	ctr, err := m.Float64ObservableCounter(i.name, i.opts...)
 	if err != nil {
-		otel.Handle(err)
+		GetErrorHandler().Handle(err)
 		return
 	}
 	i.delegate.Store(ctr)
@@ -73,7 +72,7 @@ var _ metric.Float64ObservableUpDownCounter = (*afUpDownCounter)(nil)
 func (i *afUpDownCounter) setDelegate(m metric.Meter) {
 	ctr, err := m.Float64ObservableUpDownCounter(i.name, i.opts...)
 	if err != nil {
-		otel.Handle(err)
+		GetErrorHandler().Handle(err)
 		return
 	}
 	i.delegate.Store(ctr)
@@ -102,7 +101,7 @@ var _ metric.Float64ObservableGauge = (*afGauge)(nil)
 func (i *afGauge) setDelegate(m metric.Meter) {
 	ctr, err := m.Float64ObservableGauge(i.name, i.opts...)
 	if err != nil {
-		otel.Handle(err)
+		GetErrorHandler().Handle(err)
 		return
 	}
 	i.delegate.Store(ctr)
@@ -131,7 +130,7 @@ var _ metric.Int64ObservableCounter = (*aiCounter)(nil)
 func (i *aiCounter) setDelegate(m metric.Meter) {
 	ctr, err := m.Int64ObservableCounter(i.name, i.opts...)
 	if err != nil {
-		otel.Handle(err)
+		GetErrorHandler().Handle(err)
 		return
 	}
 	i.delegate.Store(ctr)
@@ -160,7 +159,7 @@ var _ metric.Int64ObservableUpDownCounter = (*aiUpDownCounter)(nil)
 func (i *aiUpDownCounter) setDelegate(m metric.Meter) {
 	ctr, err := m.Int64ObservableUpDownCounter(i.name, i.opts...)
 	if err != nil {
-		otel.Handle(err)
+		GetErrorHandler().Handle(err)
 		return
 	}
 	i.delegate.Store(ctr)
@@ -189,7 +188,7 @@ var _ metric.Int64ObservableGauge = (*aiGauge)(nil)
 func (i *aiGauge) setDelegate(m metric.Meter) {
 	ctr, err := m.Int64ObservableGauge(i.name, i.opts...)
 	if err != nil {
-		otel.Handle(err)
+		GetErrorHandler().Handle(err)
 		return
 	}
 	i.delegate.Store(ctr)
@@ -217,7 +216,7 @@ var _ metric.Float64Counter = (*sfCounter)(nil)
 func (i *sfCounter) setDelegate(m metric.Meter) {
 	ctr, err := m.Float64Counter(i.name, i.opts...)
 	if err != nil {
-		otel.Handle(err)
+		GetErrorHandler().Handle(err)
 		return
 	}
 	i.delegate.Store(ctr)
@@ -243,7 +242,7 @@ var _ metric.Float64UpDownCounter = (*sfUpDownCounter)(nil)
 func (i *sfUpDownCounter) setDelegate(m metric.Meter) {
 	ctr, err := m.Float64UpDownCounter(i.name, i.opts...)
 	if err != nil {
-		otel.Handle(err)
+		GetErrorHandler().Handle(err)
 		return
 	}
 	i.delegate.Store(ctr)
@@ -269,7 +268,7 @@ var _ metric.Float64Histogram = (*sfHistogram)(nil)
 func (i *sfHistogram) setDelegate(m metric.Meter) {
 	ctr, err := m.Float64Histogram(i.name, i.opts...)
 	if err != nil {
-		otel.Handle(err)
+		GetErrorHandler().Handle(err)
 		return
 	}
 	i.delegate.Store(ctr)
@@ -295,7 +294,7 @@ var _ metric.Int64Counter = (*siCounter)(nil)
 func (i *siCounter) setDelegate(m metric.Meter) {
 	ctr, err := m.Int64Counter(i.name, i.opts...)
 	if err != nil {
-		otel.Handle(err)
+		GetErrorHandler().Handle(err)
 		return
 	}
 	i.delegate.Store(ctr)
@@ -321,7 +320,7 @@ var _ metric.Int64UpDownCounter = (*siUpDownCounter)(nil)
 func (i *siUpDownCounter) setDelegate(m metric.Meter) {
 	ctr, err := m.Int64UpDownCounter(i.name, i.opts...)
 	if err != nil {
-		otel.Handle(err)
+		GetErrorHandler().Handle(err)
 		return
 	}
 	i.delegate.Store(ctr)
@@ -347,7 +346,7 @@ var _ metric.Int64Histogram = (*siHistogram)(nil)
 func (i *siHistogram) setDelegate(m metric.Meter) {
 	ctr, err := m.Int64Histogram(i.name, i.opts...)
 	if err != nil {
-		otel.Handle(err)
+		GetErrorHandler().Handle(err)
 		return
 	}
 	i.delegate.Store(ctr)
